requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
nltk>=3.8.0
matplotlib>=3.7.0
PyPDF2>=3.0.0
python-docx>=0.8.11